requirements.txt
pandas
requests
plotly
folium
python-dateutil
jupyter
